+++ Tst/test_turnovers_blocking.py
@@ -4,6 +4,7 @@ from Src.start_service import start_service
 from Src.data_reposity import data_reposity
 from Src.Models.warehouse_transaction import warehouse_transaction
 from Src.Core.format_transaction import format_transaction
+from Src.Processors.turnover_process import turnover_process
 from Src.Processors.blocking_process import blocking_process
 from datetime import datetime, timedelta
 import random
@@ -20,10 +21,35 @@ class test_blocking(unittest.TestCase):
     start = start_service(reposity)
     start.create()
     
+    
+    _turnover_process = turnover_process(manager)
     _blocking_process = blocking_process(manager)
     
+    transactions = reposity.data[data_reposity.transaction_key()]
+
+
+    def test_blocking_calculations(self):
+        original_block_period = self.manager.settings.block_period
+        new_block_period = "2024-10-01"
+
+        self.manager.settings.block_period = new_block_period
+
+        # Рассчитаем оборот с новой датой блокировки
+        turnover_results_with_new_block = self._turnover_process.processing(self.transactions)
+
+        # Восстановим оригинальную дату блокировки
+        self.manager.settings.block_period = original_block_period
+
+        # Рассчитаем оборот с оригинальной датой блокировки
+        turnover_results_with_original_block = self._turnover_process.processing(self.transactions)
+        
+        assert turnover_results_with_new_block == turnover_results_with_original_block
     
-    def test_blocking(self):
+    
+    """
+    Нагрузочный тест
+    """
+    def test_load_blocking(self):
         count = 100000
         transactions = self.create_transactions(count)
         assert len(transactions) == count
@@ -40,12 +66,14 @@ class test_blocking(unittest.TestCase):
             self.manager.settings.block_period = block_date
             
             start_time = datetime.now()
-            blocked_transactions = self._blocking_process.processing(transactions)
+            self._blocking_process.processing(transactions)
             end_time = datetime.now()
 
             elapsed_time.append((end_time - start_time).total_seconds())
+        
+        
             
-        with open('results.md', 'w', encoding="utf-8") as f:
+        with open('load_test_results.md', 'w', encoding="utf-8") as f:
             f.write("# Результаты нагрузочного теста\n\n")
             f.write("| Дата блокировки | Время расчета (с) |\n")
             f.write("|------------------|-------------------|\n")
