+++ load_test_results.md
@@ -1,0 +1,7 @@
+# Результаты нагрузочного теста
+
+| Дата блокировки | Время расчета (с) |
+|------------------|-------------------|
+| 2024-02-01 | 0.0499 |
+| 2024-04-01 | 0.0443 |
+| 2024-09-01 | 0.0317 |
